+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `apiphany-integration-test` module with integration tests.
 - Added `apiphany-httpclient5` module which adds a library Apache HTTP client 5 for Apiphany.
 - Upgraded `morphix-all` to `1.0.20`.
+- Added `ApiResponse.getRequest` to return the `ApiRequest` that generated the `ApiResponse`.
 - Moved `Tests` test class to `apiphany-test` module.
 - Added `Assertions` class to `apiphany-test` module with utility JUnit style assertions.
 - Added `Assertions.assertDefaultConstructorThrows` method to assert that the default constructor throws useful for utility classes.